Added matrix utility functions to OptimMathematics.h. Implemented a system to convert screen space clicks to world space positions and to then make a ray cast to detect collisions with meshes in 3D space. For now, raycast detection is done per triangle for all meshes in the scene, will optimize later by implementing bounding volumes to mesh objects. Fixed small issue where window size is was hard coded as 1920x1080 causing issue with projection matrices that caused meshes to be stretched, and for ray cast from screen space to world space to be wrongly offset the further from the camera the object was. This made it so that the only way to get the correct projection matrix and raycast collision was to go in full screen. Now anything that needs a window's aspect ration calls a function that dynamically returns the correct width and height value.

+++ Source/Runtime/Direct3D11/Private/Dx11RHI.cpp
@@ -144,7 +144,7 @@ VertexShaderHandle Dx11RHI::createVertexShader(const char* path)
 {
   Dx11VertexShader* l_pResource = new Dx11VertexShader;
 
-  uint64 size = MultiByteToWideChar(CP_UTF8, 0, path, -1, 0, 0);
+  int32 size = MultiByteToWideChar(CP_UTF8, 0, path, -1, 0, 0);
   wchar* l_wstr = new wchar[size];
   MultiByteToWideChar(CP_UTF8, 0, path, -1, l_wstr, size);
   
@@ -159,7 +159,7 @@ VertexShaderHandle Dx11RHI::createVertexShader(const char* path)
 FragmentShaderHandle Dx11RHI::createFragmentShader(const char* path)
 {
   Dx11PixelShader* l_pResource = new Dx11PixelShader;
-  uint64 size = MultiByteToWideChar(CP_UTF8, 0, path, -1, 0, 0);
+  int32 size = MultiByteToWideChar(CP_UTF8, 0, path, -1, 0, 0);
   wchar* l_wstr = new wchar[size];
   MultiByteToWideChar(CP_UTF8, 0, path, -1, l_wstr, size);
